fix: address 23 bugs found during comprehensive codebase audit
Security fixes:
- Fix timing attack in CryptographicEquals (use FixedTimeEquals)
- Fix auth bypass via overly permissive /swagger prefix match
- Fix JSON injection in AutonomousCliSteps (escape partialId/reason)
- Fix correlation ID injection via unsanitized header values
- Fix PathSanitizer blocklist using wrong separators for Linux
- Fix path validation catching wrong exception type (PcNodeSecurityPolicy)

Reliability fixes:
- Fix middleware ordering (GlobalExceptionMiddleware must be outermost)
- Fix process stdout/stderr sequential read deadlock (AgentToolFactory)
- Fix static HttpClient disposal in QdrantSyncService
- Fix DeviceKeyAuthService race condition (add SemaphoreSlim)
- Fix unbounded iteration loop in AgentCliSteps
- Fix GlobalExceptionMiddleware missing HasStarted check
- Fix EscapeJson incomplete control character escaping (\r, \t)

Logic fixes:
- Remove read_file from dangerous tools list (non-mutating operation)
- Fix edit_file replacing all occurrences instead of first match
- Fix s.Output null check (use string.IsNullOrEmpty)
- Fix integer division truncation in SelfModelService utilization
- Fix unreachable GitHubModels validation (reorder checks)
- Add missing GitHubModels provider handler in UnifiedAIService
- Fix anthropic-version header duplication
- Fix negative history count (use Math.Clamp)
- Fix version comparison fallback (string compare is unreliable)
- Fix GitHubReleaseService header accumulation and case-sensitive content type
- Fix URL validation accepting malformed URLs (CliExecutor)

+++ src/Ouroboros.ApiHost/Services/SelfModelService.cs
@@ -54,7 +54,7 @@ public sealed class SelfModelService : ISelfModelService
                     r.Available,
                     r.Total,
                     r.Unit,
-                    Utilization = r.Total > 0 ? (r.Total - r.Available) / r.Total : 0.0
+                    Utilization = r.Total > 0 ? (double)(r.Total - r.Available) / r.Total : 0.0
                 });
 
         PerformanceDto performance = new PerformanceDto
